reorganize some types
