+++ src/util/playing-cards.ts
@@ -1,5 +1,34 @@
 import { randomInt } from 'node:crypto';
-import { type CardCode, RankCode, SuitCode, type CardColor } from '../types/cards.js';
+
+/** Valid card codes */
+export type CardCode = `${RankCode}${SuitCode}`;
+/** Valid card colors */
+export type CardColor = 'black' | 'red';
+
+/** Enum representing possible playing card ranks */
+export const enum RankCode {
+	Ace = 'A',
+	Two = '2',
+	Three = '3',
+	Four = '4',
+	Five = '5',
+	Six = '6',
+	Seven = '7',
+	Eight = '8',
+	Nine = '9',
+	Ten = '0',
+	Jack = 'J',
+	Queen = 'Q',
+	King = 'K',
+}
+
+/** Enum representing possible playing card suits */
+export const enum SuitCode {
+	Spades = 'S',
+	Clubs = 'C',
+	Hearts = 'H',
+	Diamonds = 'D',
+}
 
 /** An array of all the valid card ranks */
 export const ALL_RANKS = [
